refactor: naming
- update imports
- normalize module naming
