+++ api/_bot/commands/quote.js
@@ -1,4 +1,4 @@
 import _ from 'lodash'
-import quotes from '../data/famousQuotes'
+import quotes from '../data/json/famousQuotes'
 
 export default ctx => ctx.replyMessage(_.sample(quotes))
