+++ api/_bot/commands/solve.js
@@ -1,5 +1,6 @@
 export default ctx => {
   const args = ctx.state.command.args
+
   if (args.length < 3) {
     return ctx.replyMessage(
       'Please, provide 3 numbers a, b and c in order to' +
@@ -17,7 +18,7 @@ export default ctx => {
 
   if (a === 0 && b !== 0) {
     return ctx.replyMessage(
-      `This is a 1st grade equation ${returnString(b)}x${returnString(
+      `This is a 1st grade equation ${withSign(b)}x${withSign(
         c
       )}=0 and its solution is x=${c / b}`
     )
@@ -28,12 +29,11 @@ export default ctx => {
   }
 
   const sol = computeSolution(a, b, c)
-
   const alsoImaginarySolution = typeof sol.sol1 === 'string' ? 'imaginary' : ''
 
-  const basicMessage = `The solution to \`${returnString(a)}x²${returnString(
+  const basicMessage = `The solution to \`${withSign(a)}x²${withSign(
     b
-  )}x${returnString(c)}=0\` equation is`
+  )}x${withSign(c)}=0\` equation is`
 
   if (isDoubleSolution(sol)) {
     return ctx.replyMessage(
@@ -51,28 +51,26 @@ const computeSolution = (a, b, c) => {
     const sol = Math.sqrt(c / a)
     return { sol1: sol, sol2: -sol }
   }
+
   const sqrt = Math.sqrt(b * b - 4 * a * c)
+
   if (isNaN(sqrt)) {
     return computeImaginarySolution(a, b, c)
   }
+
   return { sol1: (-b + sqrt) / (2 * a), sol2: (-b - sqrt) / (2 * a) }
 }
 
-const returnString = solution => {
-  if (solution > 0) {
-    return `+${solution}`
-  } else {
-    return `${solution}`
-  }
-}
+const withSign = solution => (solution > 0 ? `+${solution}` : `${solution}`)
 
 const computeImaginarySolution = (a, b, c) => {
   const sqrt = Math.sqrt(Math.abs(b * b - 4 * a * c))
   const solution = { sol1: -b / (2 * a), sol2: -b / (2 * a) }
   const imaginarySolution = { sol1: sqrt / (2 * a), sol2: -sqrt / (2 * a) }
+
   return {
-    sol1: `${solution.sol1}${returnString(imaginarySolution.sol1)}i`,
-    sol2: `${solution.sol2}${returnString(imaginarySolution.sol2)}i`
+    sol1: `${solution.sol1}${withSign(imaginarySolution.sol1)}i`,
+    sol2: `${solution.sol2}${withSign(imaginarySolution.sol2)}i`
   }
 }
