+++ api/_bot/middleware/index.js
@@ -1,4 +1,4 @@
 export { default as debug } from './debug'
 export { default as error } from './error'
 export { default as replyMessage } from './reply-message'
-export { default as argParser } from './argParser'
+export { default as argParser } from './arg-parser'
