+++ api/_bot/reactions/compliment.js
@@ -1,4 +1,4 @@
 import _ from 'lodash'
-import compliments from '../data/compliments'
+import compliments from '../data/json/compliments'
 
 export default ctx => ctx.replyMessage(_.sample(compliments))
